Starting work to use an array to populate the clocks
Tried to put the array pushing into makeClocks but it would loop through the letters of the location for some reason.
Taking the array loop out of makeClocks at least draws the initial clocks correctly.
I had to disable makeClock on line 466 when pulling in parameters from the search query.
Also note that changing the cities with search does NOT repopulate clocks (which makes sense.