Changed checkQuery() function to be way more efficient & accept more incoming clocks
This is part of a larger code refactoring to allow more than 2 clocks.  checkQuery is the second 1/4 of the job (first third was to capture the clocks & times for sendURL). This code successfully updates clockAttributesArray, but not the clocks themselves. I haven't added the time toggle (&time12=true) to the URL yet and don't retrieve it with checkQuery().
3rd is to fix setTimesFromURL() to actually change the clocks.  this part will also include the 4th quarter which is converting the time using UTC.