Add a fancy icon
I chose to use the TARDIS (because you're shifting time, get it? ;-) ) If you want to use a different icon, then just grab a PNG file of whatever you want to use, and feed it to the web icon generator at https://www.favicon-generator.org/. Take the zip file that it gives you, unzip it, and shove the icons into the icons/ directory.